move cart to the separate context
