+++ src/components/header/Header.component.tsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react';
 
 import appStyles from '@/App.module.css';
 import { AppContext } from '@/context/AppContext.context.tsx';
+import { CartContext } from '@/context/Cart.context.tsx';
 import { RouterContext } from '@/context/Router.context.tsx';
 
 import { CartIconCounterComponent } from '../cart-icon-counter/CartIconCounter.component.tsx';
@@ -21,7 +22,8 @@ interface HeaderProps {
 
 export const HeaderComponent: React.FC<HeaderProps> = ({ isMobileMenuOpen, toggleMobileMenuOpen }) => {
     const { route, setRoutePath } = useContext(RouterContext);
-    const { cart, themeMode, setThemeMode } = useContext(AppContext);
+    const { cart } = useContext(CartContext);
+    const { themeMode, setThemeMode } = useContext(AppContext);
     const cartItemsCount: number = cart.items.reduce((sum, item) => sum + item.quantity, 0);
 
     const handleMobileMenuClose = (event: React.MouseEvent) => {
