+++ src/helpers/appContextHelper.ts
@@ -1,5 +1,3 @@
-import type { Cart } from '../interfaces/Cart.interface.tsx';
-import type { CartItem } from '../interfaces/CartItem.interface.tsx';
 import type { ColorScheme } from '../types/ColorScheme.type.tsx';
 
 export const parseColorScheme = (): ColorScheme => {
@@ -20,33 +18,3 @@ export const updateRootColorSchemeClass = (colorScheme: ColorScheme) => {
 };
 
 export const setStoredColorScheme = (colorScheme: ColorScheme) => localStorage.setItem('themeMode', colorScheme);
-
-export const parseStoredCartData = (): Cart => {
-    const savedCartData: string | null = getStoredCartData();
-    const parsedCartData = savedCartData ? JSON.parse(savedCartData) : null;
-
-    return isCartProps(parsedCartData) ? (parsedCartData as Cart) : { items: [] };
-};
-
-export const getStoredCartData = (): string | null => localStorage.getItem('cart');
-
-const isCartItemProps = (object: any): object is CartItem => object && 'id' in object && 'quantity' in object;
-
-const isCartProps = (object: any): object is Cart =>
-    object && Array.isArray(object.items) && object.items.every((element: any) => isCartItemProps(element));
-
-export const setStoredCartData = (cartData: Cart) => localStorage.setItem('cart', JSON.stringify(cartData));
-
-export const prepareUpdatedCartData = (currentCart: Cart, item: CartItem): Cart => {
-    const itemIndex: number = currentCart.items.findIndex((cartItem) => cartItem.id === item.id);
-
-    if (itemIndex === -1) {
-        return { items: [...currentCart.items, item] };
-    } else {
-        const newCartItems = currentCart.items.map((cartItem) =>
-            cartItem.id === item.id ? { ...cartItem, quantity: cartItem.quantity + item.quantity } : cartItem,
-        );
-
-        return { items: newCartItems };
-    }
-};
